add image_url and assets_urls on lists
As Ashley, Peter and I discussed yesterday, it would be great to have a
way to add an images and other assets to a list. Since you guys have not
really deployed piep yet, I did not add a new migration for this. Maybe
that you want to do this anyways...