add estimated duration on resources
I would like to have a way to indicate the estimated duration that a
resource will take for a student. I therefore propose to add a min and
max duration on the resources. Again, I added this to an existing
migration, but this may not be the way you want this.